Add security environment variable to run.sh script
Fix #309

Add logic to export the necessary environment variable so that executing
the 'bin/run.sh' script does not display warnings. Adding the
environment variable does not affect anything functionality, but is
rather cosmetic to eliminate the warnings for the missing environment
variable. The missing environment variable results in a warning from the
Docker engine due to the environment variable being present in the
docker-compose file which is needed for other services.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>